+++ ocrd/README.md
@@ -1,0 +1,5 @@
+# ocrd
+
+> API and CLI for ocrd
+
+See https://github.com/OCR-D/core
