+++ ocrd/bashlib/Makefile
