+++ ocrd/bashlib/README.md
