+++ ocrd/bashlib/lib.bash
