+++ ocrd/bashlib/src/dumpjson.bash
