+++ ocrd/bashlib/src/lib.bash
