+++ ocrd/bashlib/src/parse_argv.bash
