+++ ocrd/bashlib/src/usage.bash
