+++ ocrd/bashlib/src/wrap.bash
