+++ ocrd/bashlib/test/ocrd-tool.json
