+++ ocrd/ocrd/cli/bashlib.py
