+++ ocrd/ocrd/cli/ocrd_tool.py
