+++ ocrd/ocrd/lib.bash
