+++ ocrd/ocrd/workspace_bagger.py
