+++ ocrd_validators/README.md
@@ -1,0 +1,5 @@
+# ocrd_models
+
+> OCR-D framework - data validators
+
+See https://github.com/OCR-D/core
