+++ ocrd_validators/ocrd_validators/constants.py
@@ -1,0 +1,3 @@
+FILE_GROUP_PREFIX = 'OCR-D-'
+FILE_GROUP_CATEGORIES = ['IMG', 'SEG', 'OCR', 'COR', 'GT']
+IDENTIFIER_PRIORITY = ['purl', 'urn', 'doi', 'url']
