+++ ocrd_validators/requirements.txt
@@ -1,12 +1,4 @@
 bagit >= 1.7.0
 bagit_profile >= 1.3.0
 click >=7<8
-requests
-lxml
-Pillow >= 5.3.0
-numpy
-opencv-python
-Flask
 jsonschema
-pyyaml
-Deprecated == 1.2.0
