+++ ocrd_validators/setup.py
@@ -7,9 +7,9 @@ install_requires = open('requirements.txt').read().split('\n')
 install_requires.append('ocrd_utils == %s' % VERSION)
 
 setup(
-    name='ocrd_models',
+    name='ocrd_validators',
     version=VERSION,
-    description='OCR-D framework - file format APIs and schemas',
+    description='OCR-D framework - data validators',
     long_description=open('README.md').read(),
     long_description_content_type='text/markdown',
     author='Konstantin Baierer',
@@ -17,7 +17,7 @@ setup(
     url='https://github.com/OCR-D/core',
     license='Apache License 2.0',
     install_requires=install_requires,
-    packages=['ocrd_models'],
+    packages=['ocrd_validators'],
     package_data={'': ['*.json', '*.yml', '*.xml']},
     keywords=['OCR', 'OCR-D']
 )
